\name{causetext}
\alias{causetext}
\docType{data}
\title{
Translation list of COD codes
}
\description{
This is the translation of COD abbreviation codes into their corresponding full names.
}
\usage{data(causetext)}
\format{
  A data frame with the translation of COD codes to their names on 68 CODs (both the version of COD only and COD with group code). 

}

\examples{
data(causetext)
}
\keyword{datasets}
