% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA.summary.r
\name{print.interVA_summary}
\alias{print.interVA_summary}
\title{Print method for summary of the results obtained from InterVA4 algorithm}
\usage{
\method{print}{interVA_summary}(x, ...)
}
\arguments{
\item{x}{summary of InterVA results}

\item{...}{not used}
}
\description{
This function prints the summary message of the fitted results.
}
\keyword{InterVA}
