\docType{data}
\name{probbase}
\alias{probbase}
\title{Conditional probability of InterVA4}
\format{A data frame with 246 observations on 81 variables. Each observation
is the conditional probability.}
\description{
This is the table of conditional probabilities of symptoms
given CODs. The values are from InterVA-4.1.
}
\examples{
data(probbase)
}
\keyword{datasets}
