% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/InterVA4_1.7-package.R
\docType{data}
\name{probbase3}
\alias{probbase3}
\title{Conditional probability of InterVA4.03}
\format{A data frame with 246 observations on 81 variables. Each observation
is the conditional probability.}
\description{
This is the table of conditional probabilities of symptoms given CODs. The
values are from InterVA-4.03.
}
\examples{
data(probbase)
}
\keyword{datasets}

