\docType{data}
\name{SampleInput}
\alias{SampleInput}
\title{10 records of Sample Input}
\format{10 arbitrary input records.}
\description{
This is a dataset consisting of 20 arbitrary sample input
deaths in the acceptable format of InterVA4. Any data that
needs to be analyzed by this package should be in the same
format. The orders of the input fields must not be changed.
}
\examples{
data(SampleInput)
}
\keyword{datasets}
