% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA4_1.7-package.R
\docType{package}
\name{InterVA4-package}
\alias{InterVA4-package}
\alias{InterVA4}
\title{Perform InterVA4 algorithm and provide graphical summarization of COD
distribution.}
\description{
Computes individual cause of death and population cause-specific mortality
fractions using the InterVA4 algorithm. Provides a simple graphical
representation of the result.
}
\details{
To get the most up-to-date version of the package, as well as the past
versions, please check the github repository at:
\url{https://github.com/richardli/InterVA4/}

\tabular{ll}{ Package: \tab InterVA4\cr Type: \tab Package\cr Version: \tab
1.6\cr Date: \tab 2015-08-29\cr License: \tab GPL-2\cr }
}
\examples{

data(SampleInput)
sample.output <- InterVA(SampleInput, HIV = "h", Malaria = "v", write=FALSE)

}
\references{
http://www.interva.net/
}
\author{
Zehang Li, Tyler McCormick, Sam Clark

Maintainer: Zehang Li <lizehang@uw.edu>
}
\keyword{InterVA}
