\name{probbase}
\alias{probbase}
\docType{data}
\title{
Conditional probability of InterVA4
}
\description{
This is the table of conditional probabilities of symptoms given CODs. The values are from InterVA-4.1.
}
\usage{data(probbase)}
\format{
  A data frame with 246 observations on 81 variables. Each observation is the conditional probability. 
}

\examples{
data(probbase)
}
\keyword{datasets}
