% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/InterVA4_1.6-package.R
\docType{data}
\name{SampleInput}
\alias{SampleInput}
\title{10 records of Sample Input}
\format{10 arbitrary input records.}
\description{
This is a dataset consisting of 10 arbitrary sample input deaths in the
acceptable format of InterVA4. Any data that needs to be analyzed by this
package should be in the same format. The orders of the input fields must
not be changed.
}
\examples{
data(SampleInput)
}
\keyword{datasets}

